LPS-55583 Moves Java
