+++ modules/apps/user-groups-admin/user-groups-admin-test/test/integration/com/liferay/user/groups/admin/lar/test/UserGroupStagedModelDataHandlerTest.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.portlet.usergroupsadmin.lar;
+package com.liferay.user.groups.admin.lar.test;
 
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
 import com.liferay.portal.kernel.test.rule.TransactionalTestRule;
@@ -32,10 +33,12 @@ import java.util.Map;
 import org.junit.After;
 import org.junit.ClassRule;
 import org.junit.Rule;
+import org.junit.runner.RunWith;
 
 /**
  * @author David Mendez Gonzalez
  */
+@RunWith(Arquillian.class)
 public class UserGroupStagedModelDataHandlerTest
 	extends BaseStagedModelDataHandlerTestCase {
